Fix code review issues
- remove force unwraps
- replace switch setcion with
- naming
- remove redundant code
- setup logic
