+++ NewTutorials/Tutorial/AppDelegate.swift
@@ -1,4 +1,4 @@
-//  Copyright © 2016 One by Aol : Publishers. All rights reserved.
+//  Copyright © 2017 Oath. All rights reserved.
 
 import UIKit
